managed migration and seeders

+++ database/migrations/2023_09_18_073107_create_reviews_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
-            $table->integer('customer_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('customer_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->float('rating');
             $table->string('comments',1500);
             $table->foreign('customer_id')->references('id')->on('users');
